split sidebar into multiple components
