+++ lektor/admin/static/js/sidebar/Sidebar.tsx
@@ -1,28 +1,18 @@
 import React, { Component, MouseEvent } from "react";
-import { getPlatform } from "../utils";
 import { loadData } from "../fetch";
 import { trans } from "../i18n";
 import hub from "../hub";
 import { AttachmentsChangedEvent } from "../events";
 import {
   getUrlRecordPathWithAlt,
-  pathToAdminPage,
   RecordProps,
-} from "./RecordComponent";
-import Link from "./Link";
+} from "../components/RecordComponent";
+import Link from "../components/Link";
 import { bringUpDialog } from "../richPromise";
-import { Alternative, RecordInfo } from "./types";
-
-const getBrowseButtonTitle = () => {
-  const platform = getPlatform();
-  if (platform === "mac") {
-    return trans("BROWSE_FS_MAC");
-  } else if (platform === "windows") {
-    return trans("BROWSE_FS_WINDOWS");
-  } else {
-    return trans("BROWSE_FS");
-  }
-};
+import { Alternative, RecordInfo } from "../components/types";
+import PageActions from "./PageActions";
+import Alternatives from "./Alternatives";
+import AttachmentActions from "./AttachmentActions";
 
 const CHILDREN_PER_PAGE = 15;
 
@@ -77,7 +67,6 @@ class Sidebar extends Component<RecordProps, State> {
     this.childPosCache = new ChildPosCache();
 
     this.onAttachmentsChanged = this.onAttachmentsChanged.bind(this);
-    this.fsOpen = this.fsOpen.bind(this);
   }
 
   _getInitialState() {
@@ -91,7 +80,6 @@ class Sidebar extends Component<RecordProps, State> {
 
   componentDidMount() {
     this._updateRecordInfo();
-
     hub.subscribe(AttachmentsChangedEvent, this.onAttachmentsChanged);
   }
 
@@ -144,111 +132,6 @@ class Sidebar extends Component<RecordProps, State> {
           });
         }, bringUpDialog);
       }
-    );
-  }
-
-  fsOpen(event: MouseEvent) {
-    event.preventDefault();
-    loadData(
-      "/browsefs",
-      { path: this.props.record.path, alt: this.props.record.alt },
-      { method: "POST" }
-    ).then((resp) => {
-      if (!resp.okay) {
-        alert(trans("ERROR_CANNOT_BROWSE_FS"));
-      }
-    }, bringUpDialog);
-  }
-
-  renderPageActions() {
-    const urlPath = getUrlRecordPathWithAlt(
-      this.props.record.path,
-      this.props.record.alt
-    );
-
-    const { recordInfo } = this.state;
-    if (!recordInfo) {
-      return null;
-    }
-
-    const title = recordInfo.is_attachment
-      ? trans("ATTACHMENT_ACTIONS")
-      : trans("PAGE_ACTIONS");
-
-    return (
-      <div key="actions" className="section">
-        <h3>{title}</h3>
-        <ul className="nav">
-          <li key="edit">
-            <Link to={`${urlPath}/edit`}>
-              {recordInfo.is_attachment
-                ? trans("EDIT_METADATA")
-                : trans("EDIT")}
-            </Link>
-          </li>
-          {recordInfo.can_be_deleted && (
-            <li key="delete">
-              <Link to={`${urlPath}/delete`}>{trans("DELETE")}</Link>
-            </li>
-          )}
-          <li key="preview">
-            <Link to={`${urlPath}/preview`}>{trans("PREVIEW")}</Link>
-          </li>
-          {recordInfo.exists && (
-            <li key="fs-open">
-              <a href="#" onClick={this.fsOpen}>
-                {getBrowseButtonTitle()}
-              </a>
-            </li>
-          )}
-          {recordInfo.can_have_children && (
-            <li key="add-child">
-              <Link to={`${urlPath}/add-child`}>{trans("ADD_CHILD_PAGE")}</Link>
-            </li>
-          )}
-          {recordInfo.can_have_attachments && (
-            <li key="add-attachment">
-              <Link to={`${urlPath}/upload`}>{trans("ADD_ATTACHMENT")}</Link>
-            </li>
-          )}
-        </ul>
-      </div>
-    );
-  }
-
-  renderAlts() {
-    if (this.state.recordAlts.length < 2) {
-      return null;
-    }
-
-    const items = this.state.recordAlts.map((item) => {
-      let title = trans(item.name_i18n);
-      let className = "alt";
-      if (item.is_primary) {
-        title += " (" + trans("PRIMARY_ALT") + ")";
-      } else if (item.primary_overlay) {
-        title += " (" + trans("PRIMARY_OVERLAY") + ")";
-      }
-      if (!item.exists) {
-        className += " alt-missing";
-      }
-
-      const path = pathToAdminPage(
-        this.props.match.params.page,
-        getUrlRecordPathWithAlt(this.props.record.path, item.alt)
-      );
-      return (
-        <li key={item.alt} className={className}>
-          <Link to={path}>{title}</Link>
-        </li>
-      );
-    });
-
-    return (
-      <div key="alts" className="section">
-        <h3>{trans("ALTS")}</h3>
-        <ul className="nav">{items}</ul>
-      </div>
     );
   }
 
@@ -329,44 +212,26 @@ class Sidebar extends Component<RecordProps, State> {
     );
   }
 
-  renderAttachmentActions() {
-    const attachments = this.state.recordInfo?.attachments ?? [];
-    return (
-      <div key="attachments" className="section">
-        <h3>{trans("ATTACHMENTS")}</h3>
-        <ul className="nav record-attachments">
-          {attachments.length > 0 ? (
-            attachments.map((atch) => {
-              const urlPath = getUrlRecordPathWithAlt(
-                atch.path,
-                this.props.record.alt
-              );
-              return (
-                <li key={atch.id}>
-                  <Link to={`${urlPath}/edit`}>
-                    {atch.id} ({atch.type})
-                  </Link>
-                </li>
-              );
-            })
-          ) : (
-            <li key="_missing">
-              <em>{trans("NO_ATTACHMENTS")}</em>
-            </li>
-          )}
-        </ul>
-      </div>
-    );
-  }
-
   render() {
     return (
       <div className="sidebar-wrapper">
-        {this.props.record.path !== null && this.renderPageActions()}
-        {this.renderAlts()}
+        {this.props.record.path !== null && this.state.recordInfo && (
+          <PageActions
+            {...this.props}
+            recordInfo={this.state.recordInfo}
+          ></PageActions>
+        )}
+        <Alternatives
+          {...this.props}
+          recordAlts={this.state.recordAlts}
+        ></Alternatives>
         {this.state.recordInfo?.can_have_children && this.renderChildActions()}
-        {this.state.recordInfo?.can_have_attachments &&
-          this.renderAttachmentActions()}
+        {this.state.recordInfo?.can_have_attachments && (
+          <AttachmentActions
+            {...this.props}
+            recordInfo={this.state.recordInfo}
+          ></AttachmentActions>
+        )}
       </div>
     );
   }
